third_party/WebKit/LayoutTests/fast/dom/setAttributeNS-namespace-errors-expected.txt

Test that calling setAttributeNS() throws an error when http://www.w3.org/TR/dom/#dom-element-setattributens says it should.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elem.setAttributeNS(null, 'foo:bar', 'baz') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': '' is an invalid namespace for attributes..
PASS elem.setAttributeNS('http://www.w3.org/XML/1998/namespace', 'xml:abc', 'foo') did not throw exception.
PASS elem.setAttributeNS('http://www.w3.org/not-XML/1998/namespace', 'xml:abc', 'foo') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': 'http://www.w3.org/not-XML/1998/namespace' is an invalid namespace for attributes..
PASS elem.setAttributeNS('http://www.w3.org/2000/xmlns/', 'xmlns', 'http://wwww.example.org') did not throw exception.
PASS elem.setAttributeNS('http://www.w3.org/2000/not-xmlns/', 'xmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': 'http://www.w3.org/2000/not-xmlns/' is an invalid namespace for attributes..
PASS elem.setAttributeNS('http://www.w3.org/2000/xmlns/', 'xmlns:abc', 'http://wwww.example.org') did not throw exception.
PASS elem.setAttributeNS('http://www.w3.org/2000/not-xmlns/', 'xmlns:abc', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': 'http://www.w3.org/2000/not-xmlns/' is an invalid namespace for attributes..
PASS elem.setAttributeNS('http://www.w3.org/2000/xmlns/', 'badprefix:xmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': 'http://www.w3.org/2000/xmlns/' is an invalid namespace for attributes..
PASS elem.setAttributeNS('http://www.w3.org/2000/xmlns/', 'notxmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'setAttributeNS' on 'Element': 'http://www.w3.org/2000/xmlns/' is an invalid namespace for attributes..
PASS successfullyParsed is true

TEST COMPLETE
